Major code cleanup, removed by dim config
A lot of code cleanup!  See change log [2.5.0] for full details.

+++ src/main/java/net/silentchaos512/scalinghealth/network/ClientLoginMessage.java
@@ -16,7 +16,7 @@ public class ClientLoginMessage {
 
     public static ClientLoginMessage fromBytes(ByteBuf buf) {
         ClientLoginMessage msg = new ClientLoginMessage();
-        msg.areaMode = AreaDifficultyMode.fromOrdinal((int) buf.readByte());
+        msg.areaMode = AreaDifficultyMode.fromOrdinal(buf.readByte());
         msg.maxDifficultyValue = buf.readFloat();
         return msg;
     }
